ratings_api/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.13
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysql==0.0.1
MySQL-python==1.2.5
pbr==4.0.2
pytz==2018.4
requests==2.18.4
SQLAlchemy==1.2.6
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
virtualenvwrapper==4.8.2

